requirements.txt
# Core
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Export
onnx>=1.15.0
onnxruntime>=1.16.0
